webapp/requirements.txt
Flask==1.0.2
psycopg2-binary
kafka-python
requests
flask_caching